// src/api/planejador/models/PlanejamentoResponseDTO.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { ItemResDto } from "./ItemResDto";
import type { SetorResDTO } from "./SetorResDTO";

export type PlanejamentoResponseDTO = {
    /**
     * @type number
    */
    planejamentoId: number;
    /**
     * @type object
    */
    item: ItemResDto;
    /**
     * @type object
    */
    setor: SetorResDTO;
    /**
     * @type string
    */
    pedido: string;
    /**
     * @type string, date-time
    */
    dia: Date;
    /**
     * @type number
    */
    qtd: number;
};